+++ config/filenames.py
@@ -1,8 +1,22 @@
 import os.path
 
+# program data
+
 PROGRAM_DATA_DIR = "program_data"
 
 CARD_PRISONER_SAVE_FILE = os.path.join(
     PROGRAM_DATA_DIR,
     "card_prisoner.dta",
 )
+
+# assets
+
+ASSET_DIR = "assets"
+
+IMAGE_DIR = os.path.join(ASSET_DIR, "images")
+SOUND_DIR = os.path.join(ASSET_DIR, "sounds")
+BGM_DIR = os.path.join(ASSET_DIR, "bgms")
+
+# images
+
+TERRAINS = os.path.join(IMAGE_DIR, "terrains.png")